further restructuring and took SPoDictionary trait into separate file
